Implement probably-working ZFS snapshot migration
This commit adds ZFS snapshot resume support.

Now we'll build the storage volumes in two passes:

- on the first pass, we add all volumes that are empty or are ZFS-to-ZFS;

- on the second pass -- after shutting down the source -- we add all
  volumes. And because of the ZFS resume support, we sync only the
  latest changes.

This means that for a ZFS to ZFS migration, the VM only needs to go down
for a short while, while we sync the last of the data.

NOTE: For a write heavy disk we may need to do a second pass before
shutting down the VM... Otherwise downtime might be longer than we want
anyway. But right now it's at least quicker than it was.

(The slight downside to this commit is that you now get more options to
resume moves that shouldn't be done. Please resume/continue with care.)